requirements.txt
# IOK Detection Lab Requirements

# Core dependencies
selenium>=4.0.0
requests>=2.28.0
PyYAML>=6.0
urllib3>=1.26.0

# SIEM Integration
flask>=2.3.0

# Optional - for Elastic integration
elasticsearch>=8.0.0
